+++ src/configuration/env/parsers.ts
@@ -7,14 +7,16 @@ import { EnvObject, ParsedEnvObject } from './types.js';
  * @returns ParsedEnvVariables
  */
 export function parseEnvVariables(env: EnvObject): ParsedEnvObject {
-    const [REWARDS_MODE, VIP_REWARDS_MODE, SEND_UNAUTHORIZED_MESSAGE, SILENCE_MACRO_MESSAGES, ALLOW_CUSTOM_TIME_SIGNATURE, ALLOW_MANUAL_CC_MESSAGES] = getBooleanByStringList(
-        env.REWARDS_MODE,
-        env.VIP_REWARDS_MODE,
-        env.SEND_UNAUTHORIZED_MESSAGE,
-        env.SILENCE_MACRO_MESSAGES,
-        env.ALLOW_CUSTOM_TIME_SIGNATURE,
-        env.ALLOW_MANUAL_CC_MESSAGES
-    );
+    const [REWARDS_MODE, VIP_REWARDS_MODE, SEND_UNAUTHORIZED_MESSAGE, SILENCE_MACRO_MESSAGES, ALLOW_CUSTOM_TIME_SIGNATURE, ALLOW_MANUAL_CC_MESSAGES, ALLOW_MANUAL_LOOPS] =
+        getBooleanByStringList(
+            env.REWARDS_MODE,
+            env.VIP_REWARDS_MODE,
+            env.SEND_UNAUTHORIZED_MESSAGE,
+            env.SILENCE_MACRO_MESSAGES,
+            env.ALLOW_CUSTOM_TIME_SIGNATURE,
+            env.ALLOW_MANUAL_CC_MESSAGES,
+            env.ALLOW_MANUAL_LOOPS
+        );
     const TARGET_MIDI_CHANNEL = Number(env.TARGET_MIDI_CHANNEL) - 1;
 
     const TIME_SIGNATURE_NUMERATOR_CC = Number(env.TIME_SIGNATURE_NUMERATOR_CC);
@@ -32,6 +34,7 @@ export function parseEnvVariables(env: EnvObject): ParsedEnvObject {
         TIME_SIGNATURE_NUMERATOR_CC,
         TIME_SIGNATURE_DENOMINATOR_CC,
         REPETITIONS_PER_LOOP,
-        ALLOW_MANUAL_CC_MESSAGES
+        ALLOW_MANUAL_CC_MESSAGES,
+        ALLOW_MANUAL_LOOPS
     };
 }